feat: add less global variable configuration
