+++ packages/studiousFront/components.d.ts
@@ -8,9 +8,6 @@ export {}
 declare module "vue" {
   export interface GlobalComponents {
     ElButton: typeof import("element-plus/es")["ElButton"];
-    ElConfigProvider: typeof import("element-plus/es")["ElConfigProvider"];
-    ElPagination: typeof import("element-plus/es")["ElPagination"];
-    ElTable: typeof import("element-plus/es")["ElTable"];
     HelloWorld: typeof import("./src/components/HelloWorld.vue")["default"];
     IconHomeUser: typeof import("~icons/home/user")["default"];
     RouterLink: typeof import("vue-router")["RouterLink"];
