+++ packages/studiousFront/src/router/index.js
@@ -4,20 +4,24 @@
  * @version:
  * @Date: 2023-06-18 21:18:19
  * @LastEditors: CodeGetters
- * @LastEditTime: 2023-06-21 16:42:11
+ * @LastEditTime: 2023-06-22 16:33:45
  */
 import { createRouter, createWebHistory } from "vue-router";
 
 const Home = () => import("../views/HomePage.vue");
-
+const Login = () => import("../views/LoginPage.vue");
 const routes = [
   {
     path: "/",
     name: "Home",
     component: Home,
+    meta: { transition: "slide-right" },
   },
   {
-    path: "",
+    path: "/login",
+    name: "Login",
+    component: Login,
+    meta: { transition: "slide-right" },
   },
 ];
